create shell script to set environment variables
The person using this software makes a one time edit
to PROJECTDIR that sets the full file system path
to wadagen.

After that, the user can set all environment variables as follows:
cd /full/path/to/wadagen
. bin/setup.sh

This sets $PATH and $PYTHONPATH

The advantage here is that the software does not need
to be installed as a python module to work.  Not everyone
has privileges to install python modules.