add nots.py and unit tests
The program nots.py stands for "Number Of Trouble Seekers"
implying that the player characters are seeking trouble.

The larger goal here is that the player character data will
be used as input to this tool and other tools.

In this release, nots.py finds the average party level (APL) for
the set of Player Characters (PCs).  It looks at a directory of files ending
in .pcg and excludes files that represent familiars or animal companions
or mounts.

nots.py finds the Challenge Rating (CR) for the party of PCs as
the Challenge Rating may be lower for a small group of PCs, or higher
for a larger group.

nots.py may also read from a single file done up in .INI
(or configparser) format.

A set of sample data is included.  The PCs of level one were created
using PCGen 5.16

A set of unit tests is also included, to allow for refactoring and
code testing.